pyproject.toml
[build-system] 
requires = ["hatchling<=1.18.0"]
build-backend = "hatchling.build"


[options]
packages = "ppiseq"
zip_safe = "True"


[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "ppiseq"
authors = [
    {name="Hazem Essam", email="hazem@proteinea.com"},
    {name="Mohamed Elshaffei", email="shaffei@proteinea.com"},
]
maintainers = [
    {name="Hazem Essam", email="hazem@proteinea.com"},
    {name="Mohamed Elshaffei", email="shaffei@proteinea.com"},
]
version = "1.0.0"

requires-python = ">=3.7"

description = "Protein Protein Interaction research."
long_description = "file: README.md"

classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "accelerate==1.4.0",
    "esm",
    "einops==0.8.0",
    "hydra-core==1.3.2",
    "omegaconf==2.3.0",
    "scikit-learn==1.6.1",
    "torch==2.6.0",
    "datasets==3.3.1",
    "peft==0.14.0",
    "safetensors==0.4.5",
    "scikit-learn==1.6.1",
    "scipy==1.14.1",
    "sentencepiece==0.1.99",
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchvision==0.21.0",
    "torchtext==0.18.0",
    "torchmetrics==1.5.1",
    "torchelastic==0.2.2",
    "transformers==4.49.0",
    "triton==3.2.0",
    "wandb==0.18.5",
    "biopython==1.84",
]

[tool.black]
line-length = 79 
target-version = ['py37']
force-exclude = "__init__.py"